move js files into components folder
