+++ src/main/webapp/WEB-INF/jsp/overtime/overtime_form.jsp
@@ -18,8 +18,8 @@
     <script defer src="<spring:url value='/assets/npm.date-fns.5a87bdcafa645ad64030.min.js' />"></script>
     <script defer src="<spring:url value='/assets/date-fns-localized.f977abd85880a1f57b92.min.js' />"></script>
     <script defer src="<spring:url value='/assets/npm.jquery-ui.f82489a9e7eb2a4806ba.min.js' />"></script>
-    <script defer src="<spring:url value='/assets/app_form~overtime_form~sick_note_form.deaa072109f5dfc5b6d8.min.js' />"></script>
-    <script defer src="<spring:url value='/assets/overtime_form.043af4a141fe2c9930f5.min.js' />"></script>
+    <script defer src="<spring:url value='/assets/app_form~overtime_form~sick_note_form.521b6dd44d7691f214bf.min.js' />"></script>
+    <script defer src="<spring:url value='/assets/overtime_form.079aaea7b349c06092ee.min.js' />"></script>
 </head>
 <body>
