+++ src/main/webapp/WEB-INF/jsp/person/person_form.jsp
@@ -9,7 +9,7 @@
 <html>
 <head>
     <uv:head/>
-    <script defer src="<spring:url value='/assets/person_form.3dd19e5033792665a11a.min.js' />"></script>
+    <script defer src="<spring:url value='/assets/person_form.756d6ced7b4b7c0fb9cd.min.js' />"></script>
 </head>
 
 <body>
