+++ src/main/webapp/WEB-INF/jsp/sicknote/sick_note_form.jsp
@@ -25,8 +25,8 @@
     <script defer src="<spring:url value='/assets/npm.jquery-ui.f82489a9e7eb2a4806ba.min.js' />"></script>
     <script defer src="<spring:url value='/assets/npm.date-fns.5a87bdcafa645ad64030.min.js' />"></script>
     <script defer src="<spring:url value='/assets/date-fns-localized.f977abd85880a1f57b92.min' />"></script>
-    <script defer src="<spring:url value='/assets/app_form~overtime_form~sick_note_form.deaa072109f5dfc5b6d8.min.js' />"></script>
-    <script defer src="<spring:url value='/assets/sick_note_form.cf8ee78b7a527500f985.min.js' />"></script>
+    <script defer src="<spring:url value='/assets/app_form~overtime_form~sick_note_form.521b6dd44d7691f214bf.min.js' />"></script>
+    <script defer src="<spring:url value='/assets/sick_note_form.a4fa812224bfb88265b4.min.js' />"></script>
 </head>
 <body>
