+++ src/main/webapp/WEB-INF/jsp/sicknote/sick_notes.jsp
@@ -22,7 +22,7 @@
         window.uv.sickNote.person.id = "<c:out value="${sickNote.person.id}" />";
     </script>
     <script defer type="text/javascript" src="<spring:url value='/assets/npm.tablesorter.cfda1fa12e0bcfc37471.min.js' />"></script>
-    <script defer type="text/javascript" src="<spring:url value='/assets/sick_notes.bc510e7a9a91b5ad2ef7.min.js' />"></script>
+    <script defer type="text/javascript" src="<spring:url value='/assets/sick_notes.d41980c4cf47dc687d75.min.js' />"></script>
 </head>
 
 <body>
