+++ src/main/webapp/WEB-INF/jsp/workingtime/workingtime_form.jsp
@@ -10,7 +10,7 @@
 <html>
 <head>
     <uv:head/>
-    <script defer src="<spring:url value='/assets/workingtime_form.f5085ca270da56bec542.min.js' />"></script>
+    <script defer src="<spring:url value='/assets/workingtime_form.d7b6ff0b47ad83595d3a.min.js' />"></script>
 </head>
 
 <body>
