+++ src/main/webapp/WEB-INF/tags/head.tag
@@ -16,4 +16,4 @@
 <script defer src="<spring:url value='/assets/npm.bootstrap.c26a525e0f32a7cb6fc5.min.js' />"></script>
 <script defer src="<spring:url value='/assets/npm.underscore.33b5bbf6104494b62cf5.min.js' />"></script>
 <script defer src="<spring:url value='/assets/npm.font-awesome.77ab9b96300633620553.min.js' />"></script>
-<script defer src="<spring:url value='/assets/common.06ca55f4e4ab07175012.min.js' />"></script>
+<script defer src="<spring:url value='/assets/common.d1ca1050004f5e8c9616.min.js' />"></script>
