+++ src/main/webapp/bundles/account_form.js
@@ -1,1 +1,1 @@
-import '../js/back-button';
+import '../components/back-button';
