+++ src/main/webapp/bundles/app_form.js
@@ -1,10 +1,10 @@
-import '../lib/datepicker'
-import '../lib/timepicker'
+import '../components/datepicker'
+import '../components/timepicker'
 import '../lib/date-de-DE-1.0-Alpha-1'
 
 import '../js/application/app_form'
 import '../js/application/day-length-selector'
-import '../js/back-button'
+import '../components/back-button'
 
 import vacationTypeChanged from '../js/application/vacationTypeChanged'
