+++ src/main/webapp/bundles/common.js
@@ -1,12 +1,12 @@
 import $ from 'jquery';
 import '../lib/bootstrap';
 import 'font-awesome/css/font-awesome.css';
-import '../js/menu';
-import '../js/table.selectable';
-import '../js/textarea';
-import '../js/gravatar';
-import '../js/feedback';
-import tooltip from '../js/tooltip';
+import '../components/menu';
+import '../components/table-selectable';
+import '../components/textarea';
+import '../components/gravatar';
+import '../components/feedback';
+import tooltip from '../components/tooltip';
 
 global.$ = $;
 global.jQuery = $;
