+++ src/main/webapp/bundles/overtime_form.js
@@ -1,2 +1,2 @@
 import '../js/overtime/overtime_form'
-import '../js/back-button'
+import '../components/back-button'
