+++ src/main/webapp/bundles/sick_note_form.js
@@ -1,2 +1,2 @@
 import '../js/sick_notes/sick_note_form'
-import '../js/back-button'
+import '../components/back-button'
