+++ src/main/webapp/components/calendar/schraffur.png
