+++ src/main/webapp/components/datepicker/datepicker.js
@@ -1,5 +1,5 @@
 import datepicker from 'jquery-ui/ui/widgets/datepicker';
 import 'jquery-ui/themes/base/datepicker.css';
-import '../css/datepicker.css';
+import './datepicker.css';
 
 export default datepicker;
