+++ src/main/webapp/components/datepicker/index.js
@@ -1,0 +1,2 @@
+export { default } from './datepicker'
+export { default as createDatepickerInstances } from './createDatepickerInstances'
