+++ src/main/webapp/components/filter-modal/index.js
