+++ src/main/webapp/components/gravatar/index.js
