+++ src/main/webapp/components/menu/index.js
@@ -1,4 +1,4 @@
-import '../css/navbar.css'
+import './navbar.css'
 
 // UGLY FIX BECAUSE BOOTSTRAP DROPDOWN IN NAVBAR NOT WORKING ON SOME MOBILE PHONES
