+++ src/main/webapp/components/table-selectable/index.js
