+++ src/main/webapp/js/sick_notes/sick_note_form.js
@@ -1,6 +1,5 @@
 import $ from 'jquery';
-import '../../lib/datepicker';
-import createDatepickerInstances from '../createDatepickerInstances';
+import { createDatepickerInstances } from '../../components/datepicker';
 
 $(document).ready(function () {
